float-equal.c add commit

+++ float-equal.c
@@ -3,13 +3,13 @@
 #include <stdio.h>
 int main()
 {
-    float m = 1.78;
-    float x = 1.779999, y = 2;
-    float z = m / y, z1 = x / y;
+    double m = 1.78;
+    double x = 1.779999, y = 2;
+    double z = m / y, z1 = x / y;
     if (z == z1) {
         z = (m + x) / y;
     }
-    printf("Z is %f\n", z);
-    printf("Z1 is %f\n", z1);
+    printf("Z is %.9f\n", z);
+    printf("Z1 is %.9f\n", z1);
     return 0;
 }
